+++ redisson-spring/redisson-spring-cache/src/test/resources/org/redisson/spring/cache/cache-config.yaml
